generate analysis description field during import
